requirements.txt
# bcrypt==4.2.1
# blinker==1.9.0
# click==8.1.7
# colorama==0.4.6
# Deprecated==1.2.15
# dnspython==2.7.0
# Flask==3.1.0
# Flask-Limiter==3.8.0
# Flask-Login==0.6.3
# flask-marshmallow==1.2.1
# Flask-PyMongo==2.3.0
# Flask-SeaSurf==2.0.0
# Flask-SQLAlchemy==3.1.1
# Flask-WTF==1.2.2
# greenlet==3.1.1
# importlib_resources==6.4.5
# itsdangerous==2.2.0
# Jinja2==3.1.4
# joblib==1.4.2
# limits==3.13.0
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# marshmallow==3.23.1
# mdurl==0.1.2
# numpy==2.1.3
# ordered-set==4.1.0
# packaging==24.2
# Pygments==2.18.0
# pymongo==4.10.1
# python-dotenv==1.0.1
# rich==13.9.4
# scikit-learn==1.5.2
# scipy==1.14.1
# SQLAlchemy==2.0.36
# threadpoolctl==3.5.0
# typing_extensions==4.12.2
# Werkzeug==3.1.3
# wrapt==1.17.0
# WTForms==3.2.1
# cryptography
# marshmallow-sqlalchemy
# sendgrid
# # unittest-mock
# joblib
# pandas
# tensorflow
# google-generativeai

absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.15
dnspython==2.7.0
Flask==3.1.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-PyMongo==2.3.0
Flask-SeaSurf==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
gast==0.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h5py==3.12.1
httplib2==0.22.0
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.7.0
libclang==18.1.1
limits==3.13.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.2
requests==2.32.3
rich==13.9.4
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
sendgrid==6.11.0
six==1.17.0
SQLAlchemy==2.0.36
starkbank-ecdsa==2.2.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
WTForms==3.2.1
